+++ exercises/scala 3/src/traitsandinheritance/Square.scala
@@ -3,10 +3,10 @@ package traitsandinheritance
 /**
   * Created by stefanm on 22/02/2016.
   */
-class Square(val side: Int) extends Shape {
-  def sides(): Int = ???
+class Square(val side: Double) extends Shape {
+  def sides(): Double = ???
 
-  def perimeter(): Int = ???
+  def perimeter(): Double = ???
 
-  def area(): Int = ???
+  def area(): Double = ???
 }
